move events to new comp
