+++ src/App.vue
@@ -2,21 +2,7 @@
   <main class="container mx-auto my-8 space-y-8">
     <h1 class="text-4xl">Event Booking App</h1>
     <h2 class="text-2xl font-medium">All Events</h2>
-    <section class="grid grid-cols-2 gap-8">
-      <template v-if="!eventsLoading">
-        <EventCard
-          v-for="event in events"
-          :key="event.id"
-          :title="event.title"
-          :date="event.date"
-          :description="event.description"
-          @register="handleRegistration(event)"
-      /></template>
-      <template v-else>
-        <LoadingEventCard v-for="i in 4" :key="i" />
-      </template>
-    </section>
-
+    <EventList @register="handleRegistration($event)" />
     <h2 class="text-2xl font-medium">Your Bookings</h2>
     <section class="flex flex-col gap-6">
       <template v-if="!bookingLoading">
@@ -36,19 +22,10 @@
 </template>
 
 <script lang="ts">
-import EventCard from "./components/EventCard.vue";
 import BookingCard from "./components/BookingItem.vue";
-import LoadingEventCard from "./components/LoadingEventCard.vue";
 import LoadingEventBook from "./components/LoadingEventBook.vue";
+import EventList from "./components/EventList.vue";
 import { ref } from "vue";
-
-interface Event {
-  id: string;
-  title: string;
-  date: string;
-  description: string;
-  location: string;
-}
 
 type Bookings = {
   id: string;
@@ -57,41 +34,29 @@ type Bookings = {
   eventTitle: string;
   status: string;
 };
+interface Event {
+  id: string;
+  title: string;
+  date: string;
+  description: string;
+  location: string;
+}
 
 export default {
   data() {
     return {
-      events: ref<Event[]>([]),
-      eventsLoading: ref(false),
       bookings: ref<Bookings[]>([]),
       bookingLoading: ref(false),
     };
   },
 
   components: {
-    EventCard,
     BookingCard,
-    LoadingEventCard,
     LoadingEventBook,
+    EventList,
   },
 
   methods: {
-    async fetchEvents() {
-      try {
-        this.eventsLoading = true;
-        const response = await fetch("http://localhost:3001/events");
-        this.events = await response.json();
-      } catch (error) {
-        console.log("Failed to fetch events;", error);
-      } finally {
-        this.eventsLoading = false;
-      }
-    },
-
-    findBookingById(id: string): number {
-      return this.bookings.findIndex((b: Bookings) => b.id == id);
-    },
-
     async handleRegistration(event: Event) {
       if (
         this.bookings.some(
@@ -131,6 +96,9 @@ export default {
         );
       }
     },
+    findBookingById(id: string): number {
+      return this.bookings.findIndex((b: Bookings) => b.id == id);
+    },
 
     async fetchBookings() {
       try {
@@ -164,7 +132,6 @@ export default {
     },
   },
   mounted() {
-    this.fetchEvents();
     this.fetchBookings();
   },
 };
